refactor(all): move `FluentHttpClientBuilder` and `FluentHttpClientOptions` to their own files
